apm-py/flask-uwsgi-lazy-apps-logging-auto-instr/requirements.txt
backoff==1.10.0
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.42.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger-thrift==1.7.1
opentelemetry-exporter-otlp-proto-grpc==1.7.1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-flask==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
opentelemetry-propagator-b3==1.7.1
opentelemetry-proto==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
opentelemetry-util-http==0.26b1
protobuf==3.19.1
six==1.16.0
splunk-opentelemetry==1.2.0
thrift==0.15.0
uWSGI==2.0.20
Werkzeug==2.0.2
wrapt==1.13.3
